gemma3_narrative/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
aiohttp
newspaper3k
pydantic
lxml-html-clean
